+++ src/std/ntv/object.ts
@@ -1,6 +1,6 @@
 import type { Falsy, IndexSignature } from '../../typings/aliases.js';
 import type { WhenFunction } from '../../typings/comparators.js';
-import type { Entries } from '../../typings/helpers.js';
+import type { AnyFunction, Entries } from '../../typings/helpers.js';
 
 export function entries<T extends object>(o: T) {
 	return Object.entries(o) as Entries<T>;
@@ -21,12 +21,15 @@ export function freeze<T extends object>(o: T): DeepFreeze<T> {
 /** Iterate over objects while retaining its structure */
 export function iterate<T extends object>(
 	object: T,
-	fn: (entry: Entries<T>[number], index: number) => void | Falsy | [IndexSignature, any]
+	callback: AnyFunction<
+		[entry: Entries<T>[number], index: number],
+		void | Falsy | [IndexSignature, any]
+	>
 ) {
 	const pairs = entries(object);
 	const memo: typeof pairs = [];
 	for (let i = 0; i < pairs.length; i++) {
-		const res = fn(pairs[i], i);
+		const res = callback(pairs[i], i);
 		if (!res || res.length !== 2) continue;
 		memo.push(res as typeof memo[number]);
 	}