+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # mauss changelog
 
+## Unreleased
+
+- ([#174](https://github.com/alchemauss/mauss/pull/174)) add `clipboard` object to `/web` module
+
 ## 0.4.2 - 2022/11/24
 
 - ([#176](https://github.com/alchemauss/mauss/pull/176)) add empty/falsy filtering to `ntv.iterate`
 - ([#175](https://github.com/alchemauss/mauss/pull/175)) fix function in `ntv.freeze` not callable
-- ([#174](https://github.com/alchemauss/mauss/pull/174)) add `clipboard` object to `/web` module
 
 ## 0.4.1 - 2022/11/16