+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # mauss changelog
 
-## Unreleased
+## 0.4.3 - 2022/11/25
 
 - ([#177](https://github.com/alchemauss/mauss/pull/177)) fix `Entries` possibly returning `undefined`
 - ([#174](https://github.com/alchemauss/mauss/pull/174)) add `clipboard` object to `/web` module